move pages into DEPRECATED_pages folder
