+++ src/DEPRECATED_pages/registration.js
